# services/customerService/db/generated/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.27.0

package generated

type Customer struct {
	ID          int32   `json:"id"`
	Name        *string `json:"name"`
	Email       *string `json:"email"`
	Password    *string `json:"password"`
	Phonenumber *string `json:"phonenumber"`
	Address     *string `json:"address"`
}